requirements.txt
# ELF Automations Dependencies

aiofiles>=23.0.0
aiohttp>=3.9.0
alembic>=1.12.0
anthropic>=0.25.0
asyncio-mqtt>=0.13.0
asyncpg>=0.29.0  # PostgreSQL async driver
black>=23.0.0
click>=8.1.0
fastapi>=0.104.0
flake8>=6.0.0
flask>=3.0.0
httpx>=0.26.0
kubernetes>=28.0.0
mypy>=1.6.0
openai>=1.0.0
opentelemetry-api>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-sdk>=1.21.0
passlib[bcrypt]>=1.7.4
postgrest>=0.10.0
pre-commit>=3.5.0
prometheus-client>=0.19.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest>=7.4.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pyyaml>=6.0.0
redis>=5.0.0
rich>=13.0.0
sqlalchemy>=2.0.0
structlog>=23.0.0
supabase==2.9.0
tenacity>=8.2.0
uvicorn[standard]>=0.24.0

# CrewAI Agent Framework
crewai>=0.70.0

# Google A2A Protocol SDK
a2a-sdk>=0.1.0

# Vector Database
qdrant-client>=1.7.0
